requirements.txt
flask_sqlalchemy
typing
mypy-lang
flask_script
flask_migrate
